+++ tomcat/src/main/java/org/apache/coyote/exception/ContentTypeNotSupportedException.java
@@ -1,4 +1,4 @@
-package org.apache.coyote;
+package org.apache.coyote.exception;
 
 public class ContentTypeNotSupportedException extends RuntimeException {
