+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -8,10 +8,10 @@ import java.io.OutputStream;
 import java.net.Socket;
 import java.nio.charset.StandardCharsets;
 import nextstep.jwp.exception.UncheckedServletException;
-import org.apache.coyote.BodyResponse;
+import org.apache.coyote.web.BodyResponse;
 import org.apache.coyote.Processor;
-import org.apache.coyote.Request;
-import org.apache.coyote.Response;
+import org.apache.coyote.web.Request;
+import org.apache.coyote.web.Response;
 import org.apache.coyote.file.DefaultFileHandler;
 import org.apache.coyote.file.FileHandler;
 import org.apache.coyote.support.ContentType;
